Support for groups in security.
A user can be a member of a group.
Groups can have roles.
Users have the union of the roles associated with the group together
with their own roles. When logging in the user will get RolePrincipal
objects for all of these. The user will also get a GroupPrincipal object
for each group (s)he is in.
Roles are currently implemented for the PropertiesLoginModule and
PropertiesBackingEngine.
This commit also contains changes that add a JMX interceptor which can
guard JMX invocations based on user's roles.
This commit contains shell commands to manage this as well as a bunch of
new tests.